cppclass: Adding more capability to MemoryPolicy
Now memory policy also controls registration of pointers to types and
aliases for parameters and return values.  This, for example, is useful
when working with custom smart pointer implementation.

This commit also reorganizes implementation for boost::shared_ptr and
std::shared_ptr
