+++ pybindgen/wrapper_registry.py
@@ -122,9 +122,8 @@ class StdMapWrapperRegistry(WrapperRegistry):
         #code_block.write_code('std::cerr << "Erase Wrapper: obj=" <<(void *) %s << std::endl;'
         #                      % (object_rvalue))
         iterator = code_block.declare_variable("std::map<void*, PyObject*>::iterator", "wrapper_lookup_iter")
-        code_block.write_code("%(ITER)s = %(MAP)s.find((void *) %(WRAPPER)s->obj);\n"
+        code_block.write_code("%(ITER)s = %(MAP)s.find((void *) %(OBJECT_VALUE)s);\n"
                               "if (%(ITER)s != %(MAP)s.end()) {\n"
                               "    %(MAP)s.erase(%(ITER)s);\n"
                               "}\n"
-                              % dict(ITER=iterator, MAP=self.map_name, WRAPPER=wrapper_lvalue))
-
+                              % dict(ITER=iterator, MAP=self.map_name, WRAPPER=wrapper_lvalue, OBJECT_VALUE=object_rvalue))
